fix(lint): resolve staticcheck and unconvert findings in wave-2 code

+++ backend/internal/gateway/route_biz_gobd_archive.go
@@ -245,7 +245,7 @@ func parsePageParam(s string) int32 {
 		if c < '0' || c > '9' {
 			return 0
 		}
-		n = n*10 + int32(c-'0')
+		n = n*10 + (c - '0')
 	}
 	return n
 }
